rfac : improved fs and migrate to typescript
