+++ src/cli/package.json
@@ -1,26 +1,34 @@
 {
-    "name": "cli",
-    "version": "1.0.0",
-    "main": "index.js",
-    "type": "module",
-    "scripts": {
-      "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "bin": {
-      "domainforge": "./index.js"
-    },
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "description": "",
-    "dependencies": {
-      "axios": "^1.7.7",
-      "chalk": "^5.3.0",
-      "cli-table3": "^0.6.5",
-      "commander": "^12.1.0",
-      "dotenv": "^16.4.5",
-      "inquirer": "^12.0.1",
-      "node-fetch": "^3.3.2"
-    }
+  "name": "domainforge-cli",
+  "version": "1.0.2",
+  "main": "dist/index.js",
+  "type": "commonjs",
+  "scripts": {
+    "build": "tsc",
+    "start": "node dist/index.js"
+  },
+  "bin": {
+    "domainforge": "./dist/index.js"
+  },
+  "keywords": [
+    "CLI",
+    "DomainForge",
+    "domains"
+  ],
+  "author": ".mdgspace",
+  "license": "ISC",
+  "description": "It's a CLI tool to host or manage subdomains associated by the domainforge API",
+  "dependencies": {
+    "axios": "^1.7.7",
+    "chalk": "^4.1.2",
+    "cli-table3": "^0.6.5",
+    "commander": "^12.1.0",
+    "dotenv": "^16.4.5",
+    "inquirer": "^12.0.1",
+    "node-fetch": "^3.3.2"
+  },
+  "devDependencies": {
+    "@types/node": "^22.8.1",
+    "typescript": "^5.6.3"
   }
-  
+}
