+++ src/cli/tsconfig.json
@@ -1,0 +1,13 @@
+{
+    "compilerOptions": {
+      "target": "es6",
+      "module": "commonjs",
+      "outDir": "./dist",
+      "strict": true,
+      "esModuleInterop": true,
+      "skipLibCheck": true,
+      "forceConsistentCasingInFileNames": true
+    },
+    "include": ["./**/*.ts"],  // Includes all .ts files in the project.
+    "exclude": ["./node_modules", "./dist"]  // Exclude unnecessary folders.
+  }
